+++ pages/recover-password/index.tsx
