+++ store/slices/indicator-compare.ts
